adopt new escape character in WendyCom serial links
