+++ go/internal/cli/liteclient/link_direct.go
@@ -2,10 +2,13 @@ package liteclient
 
 import (
 	"bytes"
+	"crypto/rand"
 	"encoding/binary"
+	"encoding/hex"
 	"fmt"
 	"io"
 	"net"
+	"strings"
 	"sync"
 	"time"
 
@@ -14,6 +17,8 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
+const escapeChar = 0x10 // CTRL-P, aka DLE (Data Link Escape)
+
 // directLink frames WendyComMessages with the 8-byte link header over a
 // direct connection (TCP-TLS or serial). The header channel byte stays 0:
 // direct links always use the default channel.
@@ -21,6 +26,68 @@ type directLink struct {
 	conn     io.ReadWriteCloser
 	isSerial bool
 	writeMu  sync.Mutex // serializes frames across command goroutines
+}
+
+// linkHandshake switches a serial device into WendyCom mode; on TCP-TLS there
+// is nothing to set up.
+func (l *directLink) linkHandshake() error {
+	if !l.isSerial {
+		return nil
+	}
+	return serialHandshake(l.conn.(serial.Port))
+}
+
+func serialHandshake(port serial.Port) error {
+	if _, err := port.Write([]byte{escapeChar, escapeChar, escapeChar, escapeChar, 'e'}); err != nil {
+		return fmt.Errorf("serial handshake: send escape: %w", err)
+	}
+
+	var sentinel string
+
+	if err := port.SetReadTimeout(100 * time.Millisecond); err != nil {
+		return fmt.Errorf("serial handshake: set timeout: %w", err)
+	}
+
+	window := make([]byte, 0, 32)
+	oneByte := make([]byte, 1)
+	deadline := time.Now().Add(3 * time.Second)
+	for time.Now().Before(deadline) {
+		n, err := port.Read(oneByte)
+		if err != nil {
+			return fmt.Errorf("serial handshake: read: %w", err)
+		}
+		if n == 0 {
+			// rx timeout, so rx buffer empty, so send sentinel
+			var randBytes [16]byte
+			if _, err := rand.Read(randBytes[:]); err != nil {
+				return fmt.Errorf("serial handshake: generate sentinel: %w", err)
+			}
+			sentinel = hex.EncodeToString(randBytes[:])
+			window = window[:0]
+			if _, err := port.Write([]byte(strings.Repeat(" ", 16) + sentinel)); err != nil {
+				return fmt.Errorf("serial handshake: send sentinel: %w", err)
+			}
+			continue
+		}
+		if n == 1 {
+			if len(window) < 32 {
+				window = append(window, oneByte[0])
+			} else {
+				copy(window, window[1:])
+				window[31] = oneByte[0]
+			}
+			if sentinel != "" && len(window) == 32 && string(window) == sentinel {
+				if err := port.SetReadTimeout(0); err != nil {
+					return fmt.Errorf("serial handshake: clear timeout: %w", err)
+				}
+				if _, err := port.Write([]byte{escapeChar, 'm'}); err != nil {
+					return fmt.Errorf("serial handshake: send mode switch: %w", err)
+				}
+				return nil
+			}
+		}
+	}
+	return fmt.Errorf("serial handshake: sentinel not received within 3 seconds")
 }
 
 func (l *directLink) send(req *wendypb.WendyComMessage) error {
@@ -37,7 +104,7 @@ func (l *directLink) send(req *wendypb.WendyComMessage) error {
 	binary.BigEndian.PutUint16(msg[6:8], uint16(len(body)))
 	copy(msg[headerSize:], body)
 	if l.isSerial {
-		msg = bytes.ReplaceAll(msg, []byte{esc}, []byte{esc, '_'})
+		msg = bytes.ReplaceAll(msg, []byte{escapeChar}, []byte{escapeChar, '_'})
 	}
 	l.writeMu.Lock()
 	defer l.writeMu.Unlock()
@@ -73,7 +140,7 @@ func (l *directLink) maxChunk() int {
 func (l *directLink) close() error {
 	if l.isSerial {
 		if port, ok := l.conn.(serial.Port); ok {
-			_, _ = port.Write([]byte{esc, 'o'})
+			_, _ = port.Write([]byte{escapeChar, 'o'})
 			_ = port.Drain()
 		}
 	}
